ENH prefer content length over short read error
If SSL returns a "short read" when it has the content length and the
response is obviously complete due to having received all the content
based on the content length, then there is no error.

Also, try to read a header with "short read" when there is content
returned to read.  Possibly the header contains a redirect or error
information.