ENH truncate files, add named ctors
Upgrade file/update to 1.5, which includes "truncate".

1.4 added quickkey/folder path option instead of only identifying a file
by quickkey.  As this affects the potential constructors a Request can
have, we need a way to safely and correctly construct a request that has
disjoint union of requirements by the cloud.  Quickkey and file path are
both strings, so either we use named constructors, or force strong
typing for quickkeys or file paths.  Since named constructors is the
less intrusive change for users, it is more ideal to use that pattern.